requirements.txt
psycopg2-binary
pymongo
python-dotenv
rapidfuzz
openai   # optional, if you actually call the LLM
